+++ front/package.json
@@ -14,6 +14,7 @@
     "bootstrap": "^5.1.3",
     "emoji-picker-react": "^3.5.1",
     "match-sorter": "^6.3.1",
+    "bootstrap-icons": "^1.9.1",
     "react": "^18.1.0",
     "react-bootstrap": "^2.4.0",
     "react-contexify": "^5.0.0",
