+++ front/src/Components/Profile.css
@@ -1,0 +1,16 @@
+.round {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-left: calc(100% - 80px);
+  width: 50px;
+  height: 50px;
+  background: #292d32;
+  border-radius: 100%;
+  box-shadow: 2.5px 2.5px 5px rgba(31, 33, 38, 0.25), -5px -5px 15px #363c45;
+}
+
+.icon {
+  height: 20px;
+  display: inline-block;
+}
